refactor: move license validation to lib, minor cleanup
- Move licenseValidation from hooks/details/ to lib/ (not a hook)
- Extract startOfToday as module-level constant
- Add cross-reference comments between Go/TS implementations
- Use explicit undefined check for optional vehicle IDs
